+++ src/modules/api/spoonacular/spoonacular.api.service.ts
@@ -10,7 +10,7 @@ import {
 } from './spoonacular.api.types';
 import { DishIngredient, IngredientType } from '../../ingredient/ingredient.types';
 import { calculateCheckingAgain, inferDishType, inferMealType } from '../../../common/helpers';
-import { proceedIngredientUnit } from './spoonacular.api.utils';
+import { proceedIngredientUnit, proceedSummarySpaces } from './spoonacular.api.utils';
 
 @Injectable()
 export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe, SpoonacularIngredient, SpoonacularRecipeDetails, SpoonacularRecipeSections> {
@@ -77,6 +77,7 @@ export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe,
             vegetarian, vegan, glutenFree, dairyFree, veryHealthy,
             readyInMinutes, sourceName
         } = data;
+        const description = proceedSummarySpaces(summary);
         const type = inferDishType(title);
         const mealType = inferMealType(type);
 
@@ -86,7 +87,7 @@ export class SpoonacularApiService extends AbstractApiService<SpoonacularRecipe,
             ingredients: this.proceedDataToDishIngredients(extendedIngredients),
             language: 'en',
             title,
-            description: summary,
+            description,
             readyInMinutes,
             sourceOrAuthor: sourceName,
             properties: {