+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.3.0] - 2024-10-11
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Password requirements.
+- [Marek Kurańda](https://github.com/mjkuranda): Missing dish constraints.
 
 ## [2.2.1] - 2024-10-11
 ### Fixed