+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.1] - 2024-10-11
+### Fixed
+- [Marek Kurańda](https://github.com/mjkuranda): Support for other types of beverages.
+- [Marek Kurańda](https://github.com/mjkuranda): Support for singular form of units.
+
 ## [2.3.0] - 2024-10-11
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Password requirements.